# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: IncentEdge API
  version: 1.0.0
  description: |
    # IncentEdge API Documentation

    IncentEdge is an AI-powered incentive discovery, application, and monetization platform - Infrastructure's Bloomberg Terminal for Incentives.

    ## Features
    - **24,458+ verified incentive programs** across federal, state, local, and utility levels
    - **AI-powered matching** using tiered LLM system (DeepSeek + Claude)
    - **Direct Pay eligibility checking** for IRA Section 6417
    - **Multi-format report generation** (PDF, HTML, JSON)
    - **Real-time analytics** and portfolio tracking

    ## Authentication
    All authenticated endpoints require either:
    - Session-based auth (via Supabase Auth cookies)
    - API Key authentication (via `X-API-Key` header)

    ## Rate Limiting
    Rate limits vary by subscription tier:
    - **Free**: 100 requests/hour
    - **Starter**: 1,000 requests/hour
    - **Professional**: 10,000 requests/hour
    - **Team**: 50,000 requests/hour
    - **Enterprise**: Custom limits

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Handling
    The API uses standard HTTP response codes:
    - `200` - Success
    - `201` - Created
    - `400` - Bad Request (validation errors)
...
    - `404` - Not Found
...

...
    ```json
    {
...
servers:
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session token (automatic via cookies)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Validation failed
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        requestId:
          type: string
          description: Unique request identifier for support
          example: req_abc123xyz
...
    CalculateRequest:
      type: object
      required:
        - state
        - totalDevelopmentCost
        - equityInvestment
      properties:
        name:
          type: string
          description: Project name (optional)
        state:
          type: string
          minLength: 2
          maxLength: 2
          description: Two-letter state code
          example: NY
        county:
          type: string
          example: Westchester
        municipality:
          type: string
          example: Mount Vernon
        buildingType:
          type: string
          enum: [multifamily, commercial, mixed_use, industrial, residential]
          example: multifamily
        totalUnits:
          type: integer
          minimum: 0
          example: 747
        totalSqft:
          type: integer
          minimum: 0
          example: 850000
        affordableUnits:
          type: integer
          minimum: 0
          example: 224
        affordablePercentage:
          type: number
          minimum: 0
          maximum: 100
          example: 30
        totalDevelopmentCost:
          type: number
          minimum: 0
          description: Total development cost in USD
          example: 588800000
        equityInvestment:
          type: number
          minimum: 0
          description: Equity investment in USD
          example: 150000000
        projectedNOI:
          type: number
          description: Projected net operating income
        holdPeriod:
          type: integer
          minimum: 1
          maximum: 30
          default: 5
          description: Investment hold period in years
        sustainabilityTier:
          type: string
          enum: [tier_1_efficient, tier_2_high_performance, tier_3_net_zero]
        solarCapacityKw:
          type: number
          minimum: 0
          description: Solar installation capacity in kW
        storageCapacityKwh:
          type: number
          minimum: 0
          description: Battery storage capacity in kWh
        entityType:
          type: string
          enum: [corporation, partnership, llc, nonprofit, tax_exempt]
    CalculateResponse:
      type: object
      properties:
        project:
          type: object
          description: Project parameters used for calculation
        results:
          type: object
          properties:
            incentives:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  type:
                    type: string
                  category:
                    type: string
                  description:
                    type: string
                  estimatedValue:
                    type: number
                  confidence:
                    type: number
                  eligible:
                    type: boolean
                  notes:
                    type: array
                    items:
                      type: string
            totals:
              type: object
              properties:
                federal:
                  type: number
                state:
                  type: number
                local:
                  type: number
                utility:
                  type: number
                total:
                  type: number
            roiImpact:
              type: object
              properties:
                baseIRR:
                  type: number
                  description: Base case IRR without incentives (%)
                enhancedIRR:
                  type: number
                  description: Enhanced IRR with incentives (%)
                irrLift:
                  type: number
                  description: IRR improvement (percentage points)
                equityMultipleLift:
                  type: number
                  description: Equity multiple improvement
        meta:
          type: object
          properties:
            calculatedAt:
              type: string
              format: date-time
            responseTime:
              type: string
              example: 150ms
            version:
              type: string
              example: "1.0.0"
...
paths:
...
  /calculate:
    post:
      summary: Calculate Incentive Potential
      description: |
        Calculate estimated incentive values for a project based on parameters.

        Returns:
        - Matched incentive programs with estimated values
        - ROI impact analysis (IRR lift, equity multiple)
        - Confidence scores for each incentive
        - Eligibility notes and requirements

        **No authentication required** - This is a public estimation tool.
      tags:
        - Calculate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRequest'
            examples:
              multifamily:
                summary: Multifamily affordable housing project
                value:
                  state: NY
                  county: Westchester
                  buildingType: multifamily
                  totalUnits: 300
                  totalSqft: 350000
                  affordableUnits: 90
                  affordablePercentage: 30
                  totalDevelopmentCost: 250000000
                  equityInvestment: 60000000
                  sustainabilityTier: tier_2_high_performance
                  solarCapacityKw: 500
              commercial:
                summary: Commercial office building
                value:
                  state: CA
                  municipality: San Francisco
                  buildingType: commercial
                  totalSqft: 150000
                  totalDevelopmentCost: 120000000
                  equityInvestment: 30000000
                  sustainabilityTier: tier_3_net_zero
                  solarCapacityKw: 750
                  storageCapacityKwh: 1000
      responses:
        '200':
          description: Calculation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateResponse'
              examples:
                success:
                  summary: Successful calculation
                  value:
                    project:
                      state: NY
                      buildingType: multifamily
                      totalUnits: 300
                      totalSqft: 350000
                      totalDevelopmentCost: 250000000
                      equityInvestment: 60000000
                    results:
                      incentives:
                        - id: calc-45L
                          name: Section 45L Tax Credit
                          type: tax_credit
                          category: federal
                          description: "$5,000/unit for 300 units"
                          estimatedValue: 1500000
                          confidence: 0.95
                          eligible: true
                          notes:
                            - Zero Energy Ready bonus applied
                        - id: calc-179D
                          name: Section 179D Deduction
                          type: tax_deduction
                          category: federal
                          description: "$2.50/sqft for 350,000 sqft"
                          estimatedValue: 875000
                          confidence: 0.92
                          eligible: true
                          notes:
                            - Prevailing wage requirement for full deduction
                      totals:
                        federal: 45000000
                        state: 8500000
                        local: 12000000
                        utility: 450000
                        total: 65950000
                      roiImpact:
                        baseIRR: 12.5
                        enhancedIRR: 18.3
                        irrLift: 5.8
                        equityMultipleLift: 0.35
                    meta:
                      calculatedAt: "2024-02-16T12:00:00Z"
                      responseTime: 145ms
                      version: "1.0.0"
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Missing required fields: state, totalDevelopmentCost, equityInvestment"
...